+++ DebugMode/DebugModeMod.cs
@@ -2,7 +2,6 @@ using System;
 using System.Linq;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Input;
 using Pathoschild.Stardew.DebugMode.Framework;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
@@ -32,16 +31,16 @@ namespace Pathoschild.Stardew.DebugMode
         private readonly Lazy<Texture2D> Pixel = new Lazy<Texture2D>(DebugModeMod.CreatePixel);
 
         /// <summary>Keyboard keys which are mapped to a destructive action in debug mode. See <see cref="ModConfig.AllowDangerousCommands"/>.</summary>
-        private readonly Keys[] DestructiveKeys =
+        private readonly SButton[] DestructiveKeys =
         {
-            Keys.P, // ends current day
-            Keys.M, // ends current season
-            Keys.H, // randomises player's hat
-            Keys.I, // randomises player's hair
-            Keys.J, // randomises player's shirt and pants
-            Keys.L, // randomises player
-            Keys.U, // randomises farmhouse wallpaper and floors
-            Keys.F10 // tries to launch a multiplayer server and crashes
+            SButton.P, // ends current day
+            SButton.M, // ends current season
+            SButton.H, // randomises player's hat
+            SButton.I, // randomises player's hair
+            SButton.J, // randomises player's shirt and pants
+            SButton.L, // randomises player
+            SButton.U, // randomises farmhouse wallpaper and floors
+            SButton.F10 // tries to launch a multiplayer server and crashes
         };
 
 
@@ -53,15 +52,10 @@ namespace Pathoschild.Stardew.DebugMode
         public override void Entry(IModHelper helper)
         {
             // initialise
-            this.Config = helper.ReadConfig<RawModConfig>().GetParsed();
+            this.Config = helper.ReadConfig<ModConfig>();
 
             // hook events
-            ControlEvents.KeyPressed += this.ControlEvents_KeyPressed;
-            if (this.Config.Controller.HasAny())
-            {
-                ControlEvents.ControllerButtonPressed += this.ControlEvents_ControllerButtonPressed;
-                ControlEvents.ControllerTriggerPressed += this.ControlEvents_ControllerTriggerPressed;
-            }
+            InputEvents.ButtonPressed += this.InputEvents_ButtonPressed;
             LocationEvents.CurrentLocationChanged += this.LocationEvents_CurrentLocationChanged;
             GraphicsEvents.OnPostRenderEvent += this.GraphicsEvents_OnPostRenderEvent;
 
@@ -89,30 +83,18 @@ namespace Pathoschild.Stardew.DebugMode
         /// <summary>The method invoked when the player presses a keyboard button.</summary>
         /// <param name="sender">The event sender.</param>
         /// <param name="e">The event data.</param>
-        private void ControlEvents_KeyPressed(object sender, EventArgsKeyPressed e)
+        private void InputEvents_ButtonPressed(object sender, EventArgsInput e)
         {
-            // handle hotkey
-            this.HandleInput(e.KeyPressed, this.Config.Keyboard);
+            // toggle debug menu
+            if (this.Config.Controls.ToggleDebug.Contains(e.Button))
+            {
+                Program.releaseBuild = !Program.releaseBuild;
+                this.DebugMode = !this.DebugMode;
+            }
 
             // suppress dangerous actions
-            if (this.DebugMode && !this.Config.AllowDangerousCommands)
-                this.SuppressKeyIfDangerous(e.KeyPressed);
-        }
-
-        /// <summary>The method invoked when the player presses a controller button.</summary>
-        /// <param name="sender">The event sender.</param>
-        /// <param name="e">The event data.</param>
-        private void ControlEvents_ControllerButtonPressed(object sender, EventArgsControllerButtonPressed e)
-        {
-            this.HandleInput(e.ButtonPressed, this.Config.Controller);
-        }
-
-        /// <summary>The method invoked when the player presses a controller trigger button.</summary>
-        /// <param name="sender">The event sender.</param>
-        /// <param name="e">The event data.</param>
-        private void ControlEvents_ControllerTriggerPressed(object sender, EventArgsControllerTriggerPressed e)
-        {
-            this.HandleInput(e.ButtonPressed, this.Config.Controller);
+            if (this.DebugMode && !this.Config.AllowDangerousCommands && this.DestructiveKeys.Contains(e.Button))
+                e.SuppressButton();
         }
 
         /// <summary>The method invoked when the player warps into a new location.</summary>
@@ -127,34 +109,6 @@ namespace Pathoschild.Stardew.DebugMode
         /****
         ** Methods
         ****/
-        /// <summary>Suppress the specified key if it's considered dangerous (see <see cref="ModConfig.AllowDangerousCommands"/>).</summary>
-        /// <param name="key">The pressed key to suppress.</param>
-        private void SuppressKeyIfDangerous(Keys key)
-        {
-            if (this.DestructiveKeys.Contains(key))
-            {
-                Keys[] pressedKeys = Game1.oldKBState.GetPressedKeys().Union(new[] { key }).ToArray();
-                Game1.oldKBState = new KeyboardState(pressedKeys);
-            }
-        }
-
-        /// <summary>The method invoked when the player presses an input button.</summary>
-        /// <typeparam name="TKey">The input type.</typeparam>
-        /// <param name="key">The pressed input.</param>
-        /// <param name="map">The configured input mapping.</param>
-        private void HandleInput<TKey>(TKey key, InputMapConfiguration<TKey> map)
-        {
-            if (!map.IsValidKey(key))
-                return;
-
-            // perform bound action, toggle debug menu
-            if (key.Equals(map.ToggleDebug))
-            {
-                Program.releaseBuild = !Program.releaseBuild;
-                this.DebugMode = !this.DebugMode;
-            }
-        }
-
         /// <summary>Correct the player's position when they warp into an area.</summary>
         /// <param name="location">The location the player entered.</param>
         /// <param name="player">The player who just warped.</param>